mplemented the backend domain model changes and validated the build.
Summary of changes

Updated risk.go:
Changed Score to float64 and added CustomFields datatypes.JSON.
Corrected many-to-many tag to use risk_assets.
Computed score in BeforeSave as a float and simplified AfterSave history snapshot creation.
Updated history.go:
RiskHistory.Score changed from int to float64.
Updated handlers that assumed integer scores:
dashboard_handler.go: now uses float64 totals/averages and computes GlobalRiskScore appropriately.
export_handler.go: formats score with %.2f in PDF export.

+++ backend/internal/core/domain/history.go
@@ -2,22 +2,23 @@ package domain
 
 import (
 	"time"
+
 	"github.com/google/uuid"
 )
 
 // RiskHistory : Trace l'évolution d'un risque dans le temps
 type RiskHistory struct {
-	ID        uuid.UUID `gorm:"type:uuid;default:gen_random_uuid();primaryKey" json:"id"`
-	RiskID    uuid.UUID `gorm:"type:uuid;index" json:"risk_id"`
-	
+	ID     uuid.UUID `gorm:"type:uuid;default:gen_random_uuid();primaryKey" json:"id"`
+	RiskID uuid.UUID `gorm:"type:uuid;index" json:"risk_id"`
+
 	// Snapshot des valeurs clés
-	Score       int       `json:"score"`
-	Impact      int       `json:"impact"`
-	Probability int       `json:"probability"`
+	Score       float64    `json:"score"`
+	Impact      int        `json:"impact"`
+	Probability int        `json:"probability"`
 	Status      RiskStatus `json:"status"`
-	
+
 	// Qui et Quand
-	ChangedBy   string    `json:"changed_by"` // User ID ou "System" (SyncEngine)
-	ChangeType  string    `json:"change_type"` // CREATE, UPDATE, MITIGATE
-	CreatedAt   time.Time `json:"created_at"` // Timestamp du changement
+	ChangedBy  string    `json:"changed_by"`  // User ID ou "System" (SyncEngine)
+	ChangeType string    `json:"change_type"` // CREATE, UPDATE, MITIGATE
+	CreatedAt  time.Time `json:"created_at"`  // Timestamp du changement
 }
